+++ app/Console/Commands/TargetingReachFetch.php
@@ -1,13 +1,13 @@
 <?php
 /**
- * Copyright (c) 2018 Adshares sp. z o.o.
+ * Copyright (c) 2018-2019 Adshares sp. z o.o.
  *
  * This file is part of AdServer
  *
  * AdServer is free software: you can redistribute and/or modify it
  * under the terms of the GNU General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
  *
  * AdServer is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty
@@ -22,6 +22,7 @@ declare(strict_types = 1);
 namespace Adshares\Adserver\Console\Commands;
 
 use Adshares\Adserver\Console\Locker;
+use Adshares\Adserver\Http\Utils;
 use Adshares\Adserver\Models\NetworkHost;
 use Adshares\Adserver\Models\NetworkVectorsMeta;
 use Adshares\Common\Exception\RuntimeException;
@@ -31,10 +32,13 @@ use Adshares\Supply\Application\Service\SupplyClient;
 use DateTimeImmutable;
 use Illuminate\Support\Facades\DB;
 use Throwable;
+use function array_chunk;
 use function sprintf;
 
 class TargetingReachFetch extends BaseCommand
 {
+    private const CATEGORIES_CHUNK_SIZE = 200;
+
     protected $signature = 'ops:targeting-reach:fetch';
 
     protected $description = 'Fetches vectors of targeting reach and rates from remote ad servers';
@@ -129,12 +133,15 @@ class TargetingReachFetch extends BaseCommand
         DB::beginTransaction();
         try {
             DB::table('network_vectors')->where('network_host_id', $networkHostId)->delete();
-            foreach ($targetingReach['categories'] as $data) {
-                DB::table('network_vectors')->insert(
-                    [
+
+            $categoriesChunks = array_chunk($targetingReach['categories'], self::CATEGORIES_CHUNK_SIZE);
+            foreach ($categoriesChunks as $categoriesChunk) {
+                $values = [];
+                foreach ($categoriesChunk as $data) {
+                    $values[] = [
                         'network_host_id' => $networkHostId,
                         'key' => $data['key'],
-                        'data' => base64_decode($data['data']),
+                        'data' => Utils::urlSafeBase64Decode($data['data']),
                         'occurrences' => $data['occurrences'],
                         'cpm_25' => $data['cpm_25'],
                         'cpm_50' => $data['cpm_50'],
@@ -142,9 +149,11 @@ class TargetingReachFetch extends BaseCommand
                         'negation_cpm_25' => $data['negation_cpm_25'],
                         'negation_cpm_50' => $data['negation_cpm_50'],
                         'negation_cpm_75' => $data['negation_cpm_75'],
-                    ]
-                );
+                    ];
+                }
+                DB::table('network_vectors')->insert($values);
             }
+
             NetworkVectorsMeta::upsert($networkHostId, $totalEventsCount);
             DB::commit();
         } catch (Throwable $throwable) {