+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Reports stored on the server
+### Fixed
+- Publishers' domains in targeting reach
 
 ## [1.6.1] - 2020-02-20
 ### Fixed